requirements.txt
otree>=5.4.0
psycopg2>=2.8.4
sentry-sdk==0.7.9
pyjwt
cryptography
